derp integer division isn't what we want. made the same mistake with mp3 last time as well...

+++ LibAudio/WAVFormat.cs
@@ -95,7 +95,7 @@ namespace LibAudio
 
         public float GetFrameLength()
         {
-            return SamplesPerFrame/FmtHeader.SampleRate;
+            return (float)SamplesPerFrame/FmtHeader.SampleRate;
         }
 
         public void SeekToStart()
